refactor: move functions express request handlers to shared common utility
