+++ packages/gatsby/src/commands/serve.ts
@@ -7,8 +7,6 @@ import chalk from "chalk"
 import { match as reachMatch } from "@gatsbyjs/reach-router/lib/utils"
 import onExit from "signal-exit"
 import report from "gatsby-cli/lib/reporter"
-import multer from "multer"
-import cookie from "cookie"
 import telemetry from "gatsby-telemetry"
 
 import { detectPortInUseAndPrompt } from "../utils/detect-port-in-use-and-prompt"
@@ -26,6 +24,7 @@ import { reverseFixedPagePath } from "../utils/page-data"
 import { initTracer } from "../utils/tracer"
 import { configureTrailingSlash } from "../utils/express-middlewares"
 import { getDataStore, detectLmdbStore } from "../datastore"
+import { functionMiddlewares } from "../internal-plugins/functions/middleware"
 
 process.env.GATSBY_EXPERIMENTAL_LMDB_STORE = `1`
 detectLmdbStore()
@@ -168,84 +167,11 @@ module.exports = async (program: IServeProgram): Promise<void> => {
   if (functions) {
     app.use(
       `/api/*`,
-      multer().any(),
-      express.urlencoded({ extended: true }),
-      (req, _, next) => {
-        const cookies = req.headers.cookie
-
-        if (!cookies) {
-          return next()
-        }
-
-        req.cookies = cookie.parse(cookies)
-
-        return next()
-      },
-      express.text(),
-      express.json(),
-      express.raw(),
-      async (req, res, next) => {
-        const { "0": pathFragment } = req.params as { 0: string }
-
-        // Check first for exact matches.
-        let functionObj = functions.find(
-          ({ functionRoute }) => functionRoute === pathFragment
-        )
-
-        if (!functionObj) {
-          // Check if there's any matchPaths that match.
-          // We loop until we find the first match.
-          functions.some(f => {
-            if (f.matchPath) {
-              const matchResult = reachMatch(f.matchPath, pathFragment)
-              if (matchResult) {
-                req.params = matchResult.params
-                if (req.params[`*`]) {
-                  // Backwards compatability for v3
-                  // TODO remove in v5
-                  req.params[`0`] = req.params[`*`]
-                }
-                functionObj = f
-
-                return true
-              }
-            }
-
-            return false
-          })
-        }
-
-        if (functionObj) {
-          const pathToFunction = functionObj.absoluteCompiledFilePath
-          const start = Date.now()
-
-          try {
-            delete require.cache[require.resolve(pathToFunction)]
-            const fn = require(pathToFunction)
-
-            const fnToExecute = (fn && fn.default) || fn
-
-            await Promise.resolve(fnToExecute(req, res))
-          } catch (e) {
-            console.error(e)
-            // Don't send the error if that would cause another error.
-            if (!res.headersSent) {
-              res.sendStatus(500)
-            }
-          }
-
-          const end = Date.now()
-          console.log(
-            `Executed function "/api/${functionObj.functionRoute}" in ${
-              end - start
-            }ms`
-          )
-
-          return
-        } else {
-          next()
-        }
-      }
+      ...functionMiddlewares({
+        getFunctions(): Array<IGatsbyFunction> {
+          return functions
+        },
+      })
     )
   }
