+++ packages/gatsby/src/internal-plugins/functions/gatsby-node.ts
@@ -3,18 +3,15 @@ import glob from "glob"
 import path from "path"
 import webpack from "webpack"
 import _ from "lodash"
-import multer from "multer"
-import * as express from "express"
 import { getMatchPath, urlResolve } from "gatsby-core-utils"
 import { CreateDevServerArgs, ParentSpanPluginArgs } from "gatsby"
 import formatWebpackMessages from "react-dev-utils/formatWebpackMessages"
 import dotenv from "dotenv"
 import chokidar from "chokidar"
-import { match } from "@gatsbyjs/reach-router/lib/utils"
-import cookie from "cookie"
 import { reportWebpackWarnings } from "../../utils/webpack-error-utils"
 import { internalActions } from "../../redux/actions"
 import { IGatsbyFunction } from "../../redux/types"
+import { functionMiddlewares } from "./middleware"
 
 const isProductionEnv = process.env.gatsby_executing_command !== `develop`
 
@@ -478,97 +475,16 @@ export async function onCreateDevServer({
 
   app.use(
     `/api/*`,
-    multer().any(),
-    express.urlencoded({ extended: true }),
-    (req, _, next) => {
-      const cookies = req.headers.cookie
-
-      if (!cookies) {
-        return next()
-      }
-
-      req.cookies = cookie.parse(cookies)
-
-      return next()
-    },
-    express.text(),
-    express.json(),
-    express.raw(),
-    async (req, res, next) => {
-      const { "0": pathFragment } = req.params
-
-      const { functions }: { functions: Array<IGatsbyFunction> } =
-        store.getState()
-
-      // Check first for exact matches.
-      let functionObj = functions.find(
-        ({ functionRoute }) => functionRoute === pathFragment
-      )
-
-      if (!functionObj) {
-        // Check if there's any matchPaths that match.
-        // We loop until we find the first match.
-        functions.some(f => {
-          if (f.matchPath) {
-            const matchResult = match(f.matchPath, pathFragment)
-            if (matchResult) {
-              req.params = matchResult.params
-              if (req.params[`*`]) {
-                // Backwards compatability for v3
-                // TODO remove in v5
-                req.params[`0`] = req.params[`*`]
-              }
-              functionObj = f
-
-              return true
-            }
-          }
-
-          return false
-        })
-      }
-
-      if (functionObj) {
+    ...functionMiddlewares({
+      getFunctions(): Array<IGatsbyFunction> {
+        const { functions }: { functions: Array<IGatsbyFunction> } =
+          store.getState()
+        return functions
+      },
+      async prepareFn(functionObj: IGatsbyFunction): Promise<void> {
         activeDevelopmentFunctions.add(functionObj)
-
         await ensureFunctionIsCompiled(functionObj, compiledFunctionsDir)
-
-        reporter.verbose(`Running ${functionObj.functionRoute}`)
-        const start = Date.now()
-        const pathToFunction = functionObj.absoluteCompiledFilePath
-
-        try {
-          delete require.cache[require.resolve(pathToFunction)]
-          const fn = require(pathToFunction)
-
-          const fnToExecute = (fn && fn.default) || fn
-
-          await Promise.resolve(fnToExecute(req, res))
-        } catch (e) {
-          // Override the default error with something more specific.
-          if (e.message.includes(`fnToExecute is not a function`)) {
-            e.message = `${functionObj.originalAbsoluteFilePath} does not export a function.`
-          }
-          reporter.error(e)
-          // Don't send the error if that would cause another error.
-          if (!res.headersSent) {
-            res
-              .status(500)
-              .send(
-                `Error when executing function "${functionObj.originalAbsoluteFilePath}":<br /><br />${e.message}`
-              )
-          }
-        }
-
-        const end = Date.now()
-        reporter.log(
-          `Executed function "/api/${functionObj.functionRoute}" in ${
-            end - start
-          }ms`
-        )
-      } else {
-        next()
-      }
-    }
+      },
+    })
   )
 }
